+++ streamlit_canary/page.py
@@ -1,11 +1,14 @@
-import streamlit as st
 import typing as t
-from streamlit.navigation.page import StreamlitPage
+
+from ._streamlit import st
+
+if t.TYPE_CHECKING:
+    from streamlit.navigation.page import StreamlitPage
 
 
 def pages(
     elements: t.Dict[str, t.Callable[[], t.Any]]
-) -> StreamlitPage:
+) -> 'StreamlitPage':
     normalized_elements = []
     for k, v in elements.items():
         normalized_elements.append(st.Page(